requirements.txt
# API clients
groq
google-generativeai
requests

# arXiv
arxiv

# Reddit
praw

# Web framework
flask

# Environment variables
python-dotenv

# Date/time utilities
python-dateutil

# HTML generation
jinja2

# Other
flask
google-generativeai
requests
feedparser
python-dotenv
gunicorn
